WIP: Changes according to PR comments
