+++ packages/main/src/interfaces/Ui5DialogDomRef.ts
@@ -1,0 +1,6 @@
+import { Ui5DomRef } from './Ui5DomRef';
+
+export interface Ui5DialogDomRef extends Ui5DomRef {
+  open: () => void;
+  close: () => void;
+}
