+++ packages/main/src/lib/Ui5DialogDomRef.ts
@@ -1,0 +1,3 @@
+import { Ui5DialogDomRef } from '../interfaces/Ui5DialogDomRef';
+
+export { Ui5DialogDomRef };
