+++ packages/main/src/lib/Ui5DomRef.ts
@@ -1,0 +1,3 @@
+import { Ui5DomRef } from '../interfaces/Ui5DomRef';
+
+export { Ui5DomRef };
