+++ packages/main/src/lib/Ui5PopoverDomRef.ts
@@ -1,0 +1,3 @@
+import { Ui5PopoverDomRef } from '../interfaces/Ui5PopoverDomRef';
+
+export { Ui5PopoverDomRef };
